+++ taxi/commands/update.py
@@ -5,7 +5,7 @@ import six
 
 from ..aliases import Mapping
 from ..plugins import plugins_registry
-from .base import cli, populate_backends
+from .base import cli
 
 
 @cli.command(short_help="Fetch projects and shared aliases from backends.")
@@ -15,8 +15,6 @@ def update(ctx):
     Synchronizes your project database with the server and updates the shared
     aliases.
     """
-    populate_backends(ctx.obj['settings'].get_backends())
-
     ctx.obj['view'].updating_projects_database()
 
     projects = []
@@ -44,6 +42,20 @@ def update(ctx):
 
     for backend in backends_to_clear:
         ctx.obj['settings'].clear_shared_aliases(backend)
+
+    # The user can have local aliases with additional information (eg. role definition). If these aliases also exist on
+    # the remote, then they probably need to be cleared out locally to make sure they don't unintentionally use an
+    # alias with a wrong role
+    current_aliases = ctx.obj['settings'].get_aliases()
+    removed_aliases = [
+        (alias, mapping)
+        for alias, mapping in current_aliases.items()
+        if (alias in shared_aliases and shared_aliases[alias].backend == mapping.backend
+            and mapping.mapping[:2] != shared_aliases[alias].mapping[:2])
+    ]
+
+    if removed_aliases:
+        ctx.obj['settings'].remove_aliases(removed_aliases)
 
     for alias, mapping in shared_aliases.items():
         ctx.obj['settings'].add_shared_alias(alias, mapping)